wifi: mac80211: hold wiphy lock in netdev/link debugfs
It's no longer really needed to ensure that the debugfs
file isn't going away, debugfs handles that. So there's
no point in holding dev_base_lock or RTNL here, but we
should instead hold the wiphy lock since drivers will
be allowed to depend on that. Do that, which requires
splitting the sdata and link macros a bit.

Change-Id: Ic09414bfe7e6b523aca469e402877b3a85bc12af
Signed-off-by: Johannes Berg <johannes.berg@intel.com>