wifi: mac80211: lock wiphy in IP address notifier
Lock the wiphy in the IP address notifier as another
place that should have it locked before calling into
the driver. This needs a bit of attention since the
notifier can be called while the wiphy is already
locked, when we remove an interface. Handle this by
not running the notifier in this case, and instead
calling out to the driver directly.

Change-Id: Ib32eb66779542d682ad69513ae8cce3b16a0fa55
Signed-off-by: Johannes Berg <johannes.berg@intel.com>